Refactor validations (#62)
* refactor: ajust user validation return in list playlist

* refactor: ajust create playlist use case

* refactor: add validation retur in delete playlist use case

* refactor: add validation retur in details playlist use case

* refactor: add validation return in edit playlist use case

* refactor: add validation return in find playlist by id  use case

* refactor: add validation return in create playlist category use case

* refactor: add validation return in delete playlist category use case

* refactor: add validation return in edit playlist category use case

* refactor: add validation return in find playlist category use case

* refactor: add validation return in list playlist category use case

* refactor: add validation return in create scheduling use case

* refactor: add validation return in delete scheduling use case

* refactor: add validation return in edit scheduling use case

* refactor: add validation return in find scheduling use case

* refactor: add validation return in list scheduling use case

* refactor: add validation return in add playlists to scheduling use case

* refactor: add validation return in list playlists to scheduling use case

* refactor: add return validations in create content file use case

* refactor: add return validations in delete content file by id use case

* refactor: add return validations in details content file use case

* refactor: add return validations in download content file use case

* refactor: add return validations in edit content file use case

* refactor: add return validations in list content file use case

* refactor: add return validations in move file to directory use case

* refactor: add return validations in add file to playlist use case

* refactor: add return validations in delete file to playlist use case

* refactor: add return validations in find files by playlist use case

* refactor: add return validations in move file to another playlist use case

* refactor: add return validations in delete user by id use case and ajust test in move files to another playlist

* refactor: add return validations in edit user use case and ajust test in move files to another playlist

* refactor: add return validations in create auth use case

* test: add validate user test and ajust create user tests and create auth tests

* refactor: ajust providers from create and delete playlit

* test: ensure login use case

* refactor: ajust list and find user use case

+++ libs/domain/src/lib/use-case/playlist/category/delete-playlist-category.ts
@@ -1,7 +1,7 @@
 import { Inject } from '@nestjs/common';
 import { UseCase } from '../../../base/use-case';
 import { DeletePlaylistCategoryDto } from '../../../dto';
-import { EntityNotEmpty, EntityNotExists } from '../../../error';
+import { EntityNotEmpty } from '../../../error';
 import {
   DeletePlaylistCategoryRepository,
   FindPlaylistCategoryByIdRepository,
@@ -16,7 +16,7 @@ export class DeletePlaylistCategory
   constructor(
     @Inject('FindUserByIdRepository')
     private findUserByIdRepository: FindUserByIdRepository,
-    @Inject('FindUserByIdRepository')
+    @Inject('FindPlaylistCategoryByIdRepository')
     private findPlaylistCategoryByIdRepository: FindPlaylistCategoryByIdRepository,
     @Inject('DeletePlaylistCategoryRepository')
     private deletePlaylistRepository: DeletePlaylistCategoryRepository
@@ -34,12 +34,23 @@ export class DeletePlaylistCategory
       return left(new EntityNotEmpty('Logged User ID'));
     }
 
-    await ValidationUserId(loggedUserId, this.findUserByIdRepository);
+    const userValidation = await ValidationUserId(
+      loggedUserId,
+      this.findUserByIdRepository
+    );
 
-    await ValidationPlaylistCategoryId(
+    if (userValidation.isLeft()) {
+      return left(userValidation.value);
+    }
+
+    const playlistCategoryValidation = await ValidationPlaylistCategoryId(
       id,
       this.findPlaylistCategoryByIdRepository
     );
+
+    if (playlistCategoryValidation.isLeft()) {
+      return left(playlistCategoryValidation.value);
+    }
 
     await this.deletePlaylistRepository.delete(id);
