+++ libs/domain/src/lib/use-case/scheduling/delete-scheduling.ts
@@ -34,8 +34,23 @@ export class DeleteScheduling
       return left(new EntityNotEmpty('User ID'));
     }
 
-    await ValidationUserId(loggedUserId, this.findUserByIdRepository);
-    await ValidationSchedulingId(id, this.findSchedulingByIdRepository);
+    const userValidation = await ValidationUserId(
+      loggedUserId,
+      this.findUserByIdRepository
+    );
+
+    if (userValidation.isLeft()) {
+      return left(userValidation.value);
+    }
+
+    const schedulingValidation = await ValidationSchedulingId(
+      id,
+      this.findSchedulingByIdRepository
+    );
+
+    if (schedulingValidation.isLeft()) {
+      return left(schedulingValidation.value);
+    }
 
     await this.deleteSchedulingRepository.delete(input);
     return right(undefined);
